feat(ui): restructure DataGrid Header and -Toolbar stories
- create Datagrid Header stories / recipes
- trim down DataGridToolbar stories to be slim component API docs and demonstration
- add bg to DatagridToolbar for now, tokenize later
