+++ packages/ui-components/src/components/DataGridToolbar/DataGridToolbar.component.tsx
@@ -6,14 +6,14 @@
 import React, { HTMLAttributes, ReactNode } from "react"
 
 const baseToolbarStyles = `
+	jn:bg-theme-background-lvl-1
 	jn:py-3
 	jn:px-6
 	jn:mb-px
 `
 
 /**
- * `DataGridToolbar` is a spacing wrapper for a `DataGrid` header row. It provides consistent padding and separation
- * from the grid below. Use `Stack` inside to compose and position toolbar content.
+ * `DataGridToolbar` is a styled wrapper for the filter, search, and state zone (Zones 2+3) of a DataGrid header. It provides a background, consistent padding, and separation from the grid below. Use `Stack` inside to compose and position content. Zone 1 content (primary actions, bulk actions, sorting) does not use this component — use a plain `Stack` there instead.
  * @see https://cloudoperators.github.io/juno/?path=/docs/components-datagrid-datagridtoolbar--docs
  * @see {@link DataGridToolbarProps}
  */
